//bew scripts

+++ bin/release/scripts/lisSoils.py
@@ -4,8 +4,9 @@
 # author: V.G.Jetten @ 2022, 2023, 2024
 # University of Twente, Faculty ITC
 # this software has copyright model: GPLV3
+# Soilgrids layers have copyright model: https://creativecommons.org/licenses/by/4.0/
 # this software has a disclaimer
-# last edit: 17 Jan 2024
+# last edit: 11 May feb 2024
 
 from owslib.wcs import WebCoverageService
 from pcraster import *
@@ -14,6 +15,15 @@ from osgeo import gdal, gdalconst, osr, ogr
 import numpy as np
 import lisGlobals as lg
 from os.path import exists
+#import threading
+import math 
+
+SGconda = 1
+try:
+    from soilgrids import SoilGrids   
+except ImportError:    
+    SGconda = 0
+
 
 def fill_nodata(input_raster, output_raster):
     # Open the input raster dataset
@@ -30,62 +40,15 @@ def fill_nodata(input_raster, output_raster):
         return
     
     # Use gdal.FillNodata to fill nodata values
-    #gdal.FillNodata(in_ds.GetRasterBand(1), None, out_ds.GetRasterBand(1), lg.nrCols, 0)
     gdal.FillNodata(out_ds.GetRasterBand(1), None, lg.nrCols, 0)
 
     # Close datasets
     in_ds = None
     out_ds = None
+    
 
-# okay but does not work with very large datasets
-def inverse_distance_weighting(points, values, target_points, power=2, smoothing=0):
-    distances = np.linalg.norm(points - target_points[:, np.newaxis], axis=2)
-    weights = 1.0 / (distances ** power + smoothing)
-    weights /= np.sum(weights, axis=1)[:, np.newaxis]
-    interpolated_values = np.sum(weights * values, axis=1)
-    return interpolated_values
-
-def interpolate_tiff(input_tiff, output_tiff, power=2, smoothing=0):
-    gdal.UseExceptions()
-
-    src_ds = gdal.Open(input_tiff)
-    band = src_ds.GetRasterBand(1)  # Assuming a single band for simplicity
-    nodata_value = band.GetNoDataValue()
-    print(nodata_value,flush = True)
-    geotransform = src_ds.GetGeoTransform()
-    width = src_ds.RasterXSize
-    height = src_ds.RasterYSize
-
-    points = np.array([(i % width, i // width) for i in range(width * height) if band.ReadAsArray(i % width, i // width, 1, 1)[0] != nodata_value])
-    values = band.ReadAsArray()[points[:, 1], points[:, 0]]
-
-    all_points = np.array([(i % width, i // width) for i in range(width * height)])
-    all_indices = np.arange(all_points.shape[0])
-
-    missing_points = np.array([(i % width, i // width) for i in range(width * height) if band.ReadAsArray(i % width, i // width, 1, 1)[0] < 1e-5]) #== nodata_value])
-    missing_indices = np.arange(missing_points.shape[0])
-        
-    interpolated_values = inverse_distance_weighting(
-        points, values, missing_points, power=power, smoothing=smoothing
-    )
-
-    output_array = band.ReadAsArray()
-    output_array[missing_points[:, 1], missing_points[:, 0]] = interpolated_values
-
-    driver = gdal.GetDriverByName("GTiff")
-    dst_ds = driver.Create(output_tiff, width, height, 1, gdal.GDT_Float32)
-    dst_ds.SetGeoTransform(geotransform)
-    dst_ds.SetProjection(src_ds.GetProjection())
-    dst_ds.GetRasterBand(1).WriteArray(output_array)
-    dst_ds.GetRasterBand(1).SetNoDataValue(nodata_value)
-
-    src_ds = None
-    dst_ds = None
-
-
-
-class GetSoilGridsLayer:
-    "downbloading a SOILGRIDS layer from WCS service"
+class GetSoilGridsLayerConda:
+    "downbloading a SOILGRIDS layer"
     def __init__(self, x = 0):
           
         varname = lg.SG_names_[x]
@@ -95,7 +58,6 @@ class GetSoilGridsLayer:
         if lg.SG_horizon_ == 2 :
             ID = lg.SG_layers_[lg.optionSG2-1]             
             
-        self.debug = 0 #Debug_
         vname = varname  # filename for display
                
         # make a readable name for output to screen       
@@ -103,42 +65,70 @@ class GetSoilGridsLayer:
         if x == 4: vname='Gravel' 
         if x == 5: vname='Bulk Density' 
 
-        print("   => Downloading SOILGRIDS layer "+varname+ID+" as LISEM soil layer "+str(lg.SG_horizon_)+": "+vname+" content", flush=True)
+        print("   => Downloading "+varname+ID+" as LISEM soil layer "+str(lg.SG_horizon_)+": "+vname+" content", flush=True)
 
-        ESPGs = 'urn:ogc:def:crs:EPSG::{0}'.format(lg.ESPG)
-
-        if self.debug == 1:
-            print("Open SOILGRIDS WCS: "+varname+ID, flush=True)
-
-        url = "http://maps.isric.org/mapserv?map=/map/{}.map".format(varname)
-        wcs = WebCoverageService(url, version='1.0.0')
-        if self.debug == 1:
-            cov_list = list(wcs.contents)
-            mean_covs = [k for k in wcs.contents.keys() if k.find("mean") != -1]
-            print(mean_covs, flush = True)
-
-        variable = varname+ID
+        mean_covs = varname+ID
         varout = varname+str(lg.SG_horizon_)
+        
         outputnametif = "{0}.tif".format(varout)
         temptif = '_temp.tif'
-
-        dx1 = 250
-        dy1 = 250
-
-        # make a tif but the nrrows and nrcols are related to 250m now
-        # force maskbox but this is not exact, add one dx1,dy1 because sometimes the box is smaller than the DEM because of rounding off
-        Maskbox = lg.maskbox #[lg.maskbox[0]-dx1,lg.maskbox[1]-dy1,lg.maskbox[2]+2*dx1,lg.maskbox[3]+2*dy1]
-        # get data as temp geotif and save to disk
-        response = wcs.getCoverage(identifier=variable,crs=ESPGs,bbox=Maskbox,resx=dx1,resy=dy1,format='GEOTIFF_INT16')
-        with open(temptif, 'wb') as file:
-             file.write(response.read())
-
-        src = gdal.Warp(outputnametif, temptif, srcNodata = 0, xRes=dx1, yRes=dy1, outputBounds=Maskbox, resampleAlg = 'near', outputType = gdal.GDT_Float64 )     
+        if exists(temptif) :
+            os.remove(temptif)
         
-        os.remove(temptif)
-        src = None
+        # convert the mask box from the project EPSG to EPSG::152160
+        
+        # make a polygon of the maskbox coordinates
+        bbox = lg.maskbox
+        ring = ogr.Geometry(ogr.wkbLinearRing)
+        ring.AddPoint_2D(bbox[0], bbox[1])  # lower left
+        ring.AddPoint_2D(bbox[2], bbox[1])  # lower right
+        ring.AddPoint_2D(bbox[2], bbox[3])  # upper right
+        ring.AddPoint_2D(bbox[0], bbox[3])  # upper left
+        ring.AddPoint_2D(bbox[0], bbox[1])  # close the ring
+        bbox_geom = ogr.Geometry(ogr.wkbPolygon)
+        bbox_geom.AddGeometry(ring)
+        
+        # Define the source and target spatial reference systems
+        src_sr = osr.SpatialReference()
+        src_sr.ImportFromEPSG(lg.EPSG)
+        target_sr = osr.SpatialReference()
+        target_sr.ImportFromProj4('+proj=igh +datum=WGS84 +no_defs +towgs84=0,0,0')
+        # transform the box
+        transform = osr.CoordinateTransformation(src_sr, target_sr)
+        bbox_geom.Transform(transform)
+        transformed_bbox = bbox_geom.GetEnvelope()
+        # expand by one cell
+        x_min = transformed_bbox[0]-250
+        x_max = transformed_bbox[1]+250
+        y_min = transformed_bbox[2]-250
+        y_max = transformed_bbox[3]+250
+        Maskbox = [x_min,y_min,x_max,y_max]
+
+        EPSGs = 'urn:ogc:def:crs:EPSG::152160'
+       
+        if SGconda == 1:
+            #print("conda",flush = True)
+            soil_grids = SoilGrids()
+            data = soil_grids.get_coverage_data(service_id=varname, coverage_id=mean_covs, west = x_min,south=y_min,east =x_max,north=y_max, crs=EPSGs, output=temptif)
+        else :            
+            url = "https://maps.isric.org/mapserv?map=/map/{0}.map".format(varname)
+            wcs = WebCoverageService(url, version='1.0.0') 
+            response = wcs.getCoverage(identifier=mean_covs,crs=EPSGs, bbox=Maskbox,resx=250,resy=250,format='GEOTIFF_INT16')               
+            with open(temptif, 'wb') as file:
+                file.write(response.read())
+         
+        input_dataset = gdal.Open(temptif)        
+        gdal.Warp(outputnametif, input_dataset, 
+            format='GTiff', 
+            srcNodata = -32768,
+            xRes=250, yRes=250, 
+            outputType = gdal.GDT_Float32)        
+        
+        # Close the datasets
+        input_dataset = None
         response = None
-        
+        if exists(temptif) :
+            os.remove(temptif)
 
 
 ### ---------- class SoilGridsTransform() ---------- ###
@@ -178,7 +168,18 @@ class SoilGridsTransform(StaticModel):
         if lg.optionResample == 2 :
             resa = 'cubic'       
             
-        src = gdal.Warp(tempwarptif,temptif, srcNodata = 0, xRes=lg.dx, yRes=lg.dy, outputBounds=lg.maskbox, resampleAlg = resa, outputType = gdal.GDT_Float64 )
+        warp_options = {
+            "xRes": lg.dx,
+            "yRes": lg.dy,
+            "outputBounds": lg.maskbox,
+            "resampleAlg": resa,
+            "outputType": gdal.GDT_Float32,
+            "dstSRS": 'EPSG:{0}'.format(lg.EPSG)
+        }
+  
+        src = gdal.Warp(tempwarptif, temptif, **warp_options)    
+            
+        #src = gdal.Warp(tempwarptif,temptif, srcNodata = 0, xRes=lg.dx, yRes=lg.dy, outputBounds=lg.maskbox, resampleAlg = resa, outputType = gdal.GDT_Float64 )
         src = None
         
         #open the tif and create PCRaster copy
